# backend/openapi.yaml
openapi: 3.0.1
info:
  title: Alumns App API (Mock)
  version: 1.0.0
  description: Mock API for local development and integration tests
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string

paths:
  /login:
    post:
      summary: Login a user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: invalid credentials

  /profile:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: invalid or expired token

  /posts:
    get:
      summary: List posts (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: paginated posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  meta:
                    type: object
                    properties:
                      page:
                        type: integer
                      total:
                        type: integer